refactor(cas): classify database failures in db, not in the HTTP layer
`db::classify` decides whether a `sqlx::Error` means the database is
unavailable (SQLSTATE classes 08, 53, 57, 58, protocol and BEGIN failures),
busy (serialization failure, deadlock), or something the code has no name
for. `http::error` only maps that verdict to 503/500, so the next transport
does not repeat the SQLSTATE table.
